feat: break dashboard into streakanalysis and activities
