+++ src/Pages/Dashboard/helpers/dashboard.helpers.ts
@@ -1,9 +1,17 @@
-export const navigateToStreakList = (navigate: any) => {
-  navigate("/streak-list");
+export const navigateToStreakList = (navigate: any, goTo: string = "") => {
+  navigate("/streak-list", { state: { goTo } });
+};
+
+export const navigateToRewardList = (navigate: any, goTo: string = "") => {
+  navigate("/reward-list", { state: { goTo } });
+};
+
+export const navigateToRecentActivities = (navigate: any) => {
+  navigate("/recent-activities");
 };
 
 export const navigateToStreak = (
-  e : any,
+  e: any,
   navigate: any,
   id: number,
   from: string
