+++ src/Styles/Pages/streak.scss
@@ -248,6 +248,7 @@
                             }
                         }
                     }
+                    
                 }
             }
         }
